Rename jws 1.0 to 1.1 everywhere.
