+++ instrumentation/jaxws/jws-1.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/jaxws/jws/v1_1/WebServiceClass.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.jaxws.jws.v1_0;
+package io.opentelemetry.javaagent.instrumentation.jaxws.jws.v1_1;
 
 import javax.jws.WebService;
